requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
mysql-connector-python
jinja2
python-multipart

email-validator
bcrypt
PyJWT
Pillow
httpx
pytest


# 新增 AWS SDK
boto3==1.35.0 